[Front]_Fix_필름스팟 페이지 변경사항 폐기

+++ server/controller/filmtalk_comments/deletion.js
@@ -1,19 +1,28 @@
 const { filmtalk_comments } = require("../../models");
 
 module.exports = {
+  // 필름토크 댓글 삭제 기능
   delete: async (req, res) => {
     try {
       const { filmtalk_comments_id } = req.params;
 
-      await filmtalk_comments.destroy({
+      const userInfo = await filmtalk_comments.findOne({
         where: {
-          id: filmtalk_comments_id,
-        },
-      });
-
-      res.status(200).send({
-        message: "Successfully deleted",
-      });
+          id: filmtalk_comments_id
+        }
+      })
+      if (!userInfo) {
+        res.status(404).send({ message: "Failed to load information" })
+      } else {
+        await filmtalk_comments.destroy({
+          where: {
+            id: filmtalk_comments_id,
+          },
+        });
+        res.status(200).send({
+          message: "Successfully deleted",
+        });
+      }
     } catch (err) {
       console.error(err);
       res.status(500).send({
